cicd/jenkins_script/python3_script/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Deprecated==1.2.11
google-auth==1.27.1
idna==2.10
kubernetes==12.0.1
multi-key-dict==2.0.3
oauthlib==3.1.0
pathspec==0.8.1
pbr==5.5.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyGithub==1.54.1
PyJWT==1.7.1
python-dateutil==2.8.1
python-jenkins==1.7.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
urllib3==1.26.3
websocket-client==0.58.0
wrapt==1.12.1
